CORE-18698: move FilteredTransaction interface and its data classes to ledger-common-flow
