+++ libs/ledger/ledger-common-data/src/main/kotlin/net/corda/ledger/common/data/transaction/filtered/FilteredTransaction.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.common.flow.transaction.filtered
+package net.corda.ledger.common.data.transaction.filtered
 
 import net.corda.ledger.common.data.transaction.WireTransaction
 import net.corda.v5.base.annotations.CordaSerializable
